Revise UI for adding user to organization
